requirements.txt
streamlit
supabase-py
sendgrid